+++ test/print.cpp
@@ -28,6 +28,7 @@
 #include <sstream>
 
 #include <kitty/print.hpp>
+#include <kitty/properties.hpp>
 #include <kitty/static_truth_table.hpp>
 
 #include "utility.hpp"
@@ -76,4 +77,25 @@ TEST_F( PrintTest, print_raw )
 
     EXPECT_EQ( tt, tt2 );
   }
+}
+
+TEST_F( PrintTest, xmas_one )
+{
+  std::stringstream ss;
+  static_truth_table<5> maj;
+  create_majority( maj );
+  print_xmas_tree_for_function( maj, ss );
+
+  EXPECT_EQ( 574u, ss.str().size() );
+}
+
+TEST_F( PrintTest, xmas_all )
+{
+  std::stringstream ss;
+  print_xmas_tree_for_functions<dynamic_truth_table>( 3,
+                                                      {{is_krom<dynamic_truth_table>, {31}},
+                                                       {is_horn<dynamic_truth_table>, {32}},
+                                                       {[]( const dynamic_truth_table& f ) { return is_horn( f ) && is_krom( f ); }, {33}}},
+                                                      ss );
+  EXPECT_EQ( 7046u, ss.str().size() );
 }